Server/requirements.txt
blinker      ~=1.7.0
click        ~= 8.1.7
colorama     ~= 0.4.6
Flask         ~=3.0.2
grpcio       ~= 1.62.1
grpcio-tools  ~=1.62.1
itsdangerous  ~=2.1.2
Jinja2        ~=3.1.3
MarkupSafe    ~=2.1.5
pip           ~=23.2.1
protobuf     ~= 4.25.3
setuptools   ~= 69.1.1
Werkzeug     ~= 3.0.1
